refactor(livestock): split livestock-kind-page into focused modules
Extract SearchBar, LivestockUnitRow, and useLivestockAssetEdit out of
the 435-line page file to satisfy the 200-line and SRP rules.
